requirements.txt
instagrapi>=2.1.2
groq>=0.4.0
python-dotenv>=1.0.0
